[Minjin.Kim] [CON-68] 프로젝트 공고 생성 API 연결 - 프로젝트 공고 생성 API 연결 - 리프래시 토큰 갱신 로직 변경

+++ presentation/src/main/java/com/connectcrew/presentation/model/project/ProjectFeedLikeInfo.kt
@@ -3,7 +3,7 @@ package com.connectcrew.presentation.model.project
 import com.connectcrew.domain.usecase.project.entity.ProjectFeedLikeInfoEntity
 
 data class ProjectFeedLikeInfo(
-    val projectId: Int,
+    val projectId: Long,
     val isLike: Boolean,
     val likeCount: Int
 )
